backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9
motor>=3.4
openai>=1.0
elevenlabs>=1.0
python-dotenv>=1.0
httpx>=0.27
